+++ lib/change_health/request/trading_partner.rb
@@ -11,7 +11,7 @@ module ChangeHealth
 
         response = ChangeHealth::Connection.new.request(endpoint: ENDPOINT, verb: :get, query: params)
         trading_partners_data = ChangeHealth::Response::TradingPartnersData.new(response['payers'])
-        trading_partners_data.map { |partner| ChangeHealth::Models::TradingPartner.new(name: partner.name, service_id: partner.service_id) }
+        trading_partners_data.map { |partner| ChangeHealth::Models::TradingPartner.new(name: partner.name, alias: partner.alias, line_of_business: partner.line_of_business, plan_type: partner.plan_type, service_id: partner.service_id )}
       end
     end
   end