+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+# [5.0.1] - 2023-08-31
+
+### Added
+
+* added alias to trading_partner_data
+* added line_of_business to trading_partner_data
+* added plan_type to trading_partner_data
+
 # [5.0.0] - 2022-08-24
 
 ### Changed
@@ -551,6 +559,7 @@ Added the ability to hit professional claim submission API. For more details, se
 * Authentication
 * Configuration
 
+[5.0.1]: https://github.com/WeInfuse/change_health/compare/v5.0.0...v5.0.1
 [5.0.0]: https://github.com/WeInfuse/change_health/compare/v4.14.0...v5.0.0
 [4.14.0]: https://github.com/WeInfuse/change_health/compare/v4.13.1...v4.14.0
 [4.13.1]: https://github.com/WeInfuse/change_health/compare/v4.13.0...v4.13.1